yenc by section instead of by line
Instead of inserting the escape characters after the line is encoded,
split the original line at the positions where escape characters would
fall and encode each section individually.  This addresses the
performance issue of the previous version which was caused by creating
and destroying numerous string literals in a loop.